Move ChemblClient into clients package
