+++ library/chembl_document.py
@@ -7,7 +7,7 @@ from typing import Any
 
 import pandas as pd
 
-from .chembl_client import ChemblClient, _chunked
+from library.clients import ChemblClient, _chunked
 from .config import ApiCfg
 
 DOCUMENT_COLUMNS = [
