+++ scripts/get_activity_data.py
@@ -13,7 +13,7 @@ from pandera.errors import SchemaErrors
 from library import chembl_library as cl
 from library import cli
 from library import io
-from library.chembl_client import ChemblClient
+from library.clients import ChemblClient
 from library.cli import (
     LoggerConfig,
     configure_logger,
