+++ scripts/get_testitem_data.py
@@ -29,7 +29,7 @@ from library import chembl_library as cl
 from library import cli
 from library import io, molecule_catalog, testitem_enrichment
 from library import pubchem_library as pl
-from library.chembl_client import ChemblClient
+from library.clients import ChemblClient
 from library.cli import (
     LoggerConfig,
     configure_logger,
