+++ tests/test_activity_extraction.py
@@ -12,7 +12,7 @@ from library.config import Config
 
 
 class DummyChemblClient:
-    """Minimal stub replacing :class:`library.chembl_client.ChemblClient`."""
+    """Minimal stub replacing :class:`library.clients.ChemblClient`."""
 
     def __init__(self, *args, **kwargs) -> None:  # noqa: D401 - signature mirrors context
         self.closed = False
